Add include-user-parameters-in-signature-p parameter
Twitter wants the OAuth signature to include only authentication
related parameters.  This commit adds a
INCLUDE-USER-PARAMETERS-IN-SIGNATURE-P keyword parameter to
OBTAIN-REQUEST-TOKEN and ACCESS-PROTECTED-RESOURCE which can be passed
as NIL to exclude non-authentication related parameters from
calculating the signature.  By default, the parameter is passed as T
to obtain the previous behavior.